+++ modules/vehicle/__init__.py
@@ -1,0 +1,4 @@
+from modules.vehicle.manager import VehicleManager
+from modules.vehicle.models import VehicleStatus
+
+__all__ = ["VehicleManager", "VehicleStatus"]